Updated according to comments
