+++ amadeus/Shopping.cs
@@ -45,6 +45,18 @@ namespace amadeus
 
         /// <summary>
         /// A namespaced client for the
+        /// <code>/v2/shopping/flight-offers</code> endpoints.
+        /// </summary>
+        public FlightOffersSearch flightOffersSearch;
+
+        /// <summary>
+        /// A namespaced client for the
+        /// <code>/v1/shopping/seatmaps</code> endpoints.
+        /// </summary>
+        public SeatMapDisplay seatMapDisplay;
+
+        /// <summary>
+        /// A namespaced client for the
         /// <code>/v2/shopping/hotel-offers</code> endpoints.
         /// </summary>
         ///
@@ -66,6 +78,8 @@ namespace amadeus
             this.flightDates = new FlightDates(client);
             this.flightDestinations = new FlightDestinations(client);
             this.flightOffers = new FlightOffers(client);
+            this.flightOffersSearch = new FlightOffersSearch(client);
+            this.seatMapDisplay = new SeatMapDisplay(client);
             this.hotelOffers = new HotelOffers(client);
             this.hotelOffersByHotel = new HotelOffersByHotel(client);
         }
